app/requirements.txt
 Django>=2.0,<3.0
 psycopg2-binary
 psycopg2
 djangorestframework
 markdown
 django-filter